+++ CHANGELOG.md
@@ -7,3 +7,6 @@ All notable changes to this project will be documented in this file
 15/1/25
 - Created CHANGELOG.md
 - Created an intro splash screen to the app, to be improved later
+
+16/1/25
+- Created the main menu, featuring two buttons, one to start a trip and one to view past trips.